pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "usb-explorer"
version = "0.1.0"
description = "Visual USB topology and debugging for Linux"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
keywords = ["usb", "linux", "debugging", "hardware", "monitoring", "topology"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware",
    "Topic :: System :: Monitoring",
    "Topic :: Utilities",
]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "python-socketio[asyncio]>=5.0.0",
    "pyudev>=0.24.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/usb-explorer"
Repository = "https://github.com/yourusername/usb-explorer"
Issues = "https://github.com/yourusername/usb-explorer/issues"

[project.scripts]
usb-explorer = "usb_explorer.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
usb_explorer = [
    "templates/*.html",
    "static/css/*.css",
    "static/js/*.js",
    "static/img/*",
    "static/sounds/*",
]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
